feat: remove styled-components, styled-system; update components (removing injected theme provider)
