+++ apps/web/src/pages/Contracts/components/LiquiditySafeContractsTable.tsx
@@ -13,8 +13,6 @@ import {
 } from '@mui/material';
 import KeyboardArrowDownIcon from '@mui/icons-material/KeyboardArrowDown';
 import KeyboardArrowUpIcon from '@mui/icons-material/KeyboardArrowUp';
-import { ThemeProvider } from '@mui/material/styles';
-import theme from '../../../theme_mui';
 
 import CopyButton from './CopyButton';
 import arbitrumLogo from '../../../img/arbitrum-shield.png';
@@ -45,66 +43,64 @@ const LiquiditySafeContractsTable: React.FC = () => {
   const handleCopyToClipboard = (address: string) => navigator.clipboard.writeText(address);
 
   return (
-    <ThemeProvider theme={theme}>
-      <TableContainer component={Paper}>
-        <Table aria-label="liquidity safe table">
-          <TableHead>
-            <TableRow>
-              <TableCell style={{ width: '35%' }}>Name</TableCell>
-              <TableCell style={{ width: '50%' }}>Address</TableCell>
-              <TableCell style={{ width: '15%' }}>Actions</TableCell>
-            </TableRow>
-          </TableHead>
-          <TableBody>
-            <TableRow onClick={handleToggle} style={{ cursor: 'pointer' }}>
-              <TableCell colSpan={3} style={{ textAlign: 'center' }}>
-                {open ? ' Hide Details' : ' Show Details'}
-                <IconButton size="small">
-                  {open ? <KeyboardArrowUpIcon /> : <KeyboardArrowDownIcon />}
-                </IconButton>
+    <TableContainer component={Paper}>
+      <Table aria-label="liquidity safe table">
+        <TableHead>
+          <TableRow>
+            <TableCell style={{ width: '35%' }}>Name</TableCell>
+            <TableCell style={{ width: '50%' }}>Address</TableCell>
+            <TableCell style={{ width: '15%' }}>Actions</TableCell>
+          </TableRow>
+        </TableHead>
+        <TableBody>
+          <TableRow onClick={handleToggle} style={{ cursor: 'pointer' }}>
+            <TableCell colSpan={3} style={{ textAlign: 'center' }}>
+              {open ? ' Hide Details' : ' Show Details'}
+              <IconButton size="small">
+                {open ? <KeyboardArrowUpIcon /> : <KeyboardArrowDownIcon />}
+              </IconButton>
+            </TableCell>
+          </TableRow>
+          {open && liquiditySafeAddresses.map((wallet) => (
+            <TableRow key={wallet.address}>
+              <TableCell>
+                <Box sx={{ display: 'flex' }}>
+                  <Typography variant="body2">
+                    {wallet.name}
+                  </Typography>
+                </Box>
+              </TableCell>
+              <TableCell>
+                <Box sx={{ display: 'flex' }}>
+                  <Typography variant="body2" noWrap>
+                    {wallet.address}
+                  </Typography>
+                  <CopyButton onClick={() => handleCopyToClipboard(wallet.address)}>
+                    Copy
+                  </CopyButton>
+                </Box>
+              </TableCell>
+              <TableCell>
+                <Box>
+                  <a
+                    href={wallet.arbiscanLink}
+                    target="_blank"
+                    rel="noopener noreferrer"
+                    style={{ cursor: 'pointer', marginRight: 10 }}
+                  >
+                    <img
+                      src={arbitrumLogo}
+                      alt="arbitrum icon"
+                      width={30}
+                    />
+                  </a>
+                </Box>
               </TableCell>
             </TableRow>
-            {open && liquiditySafeAddresses.map((wallet) => (
-              <TableRow key={wallet.address}>
-                <TableCell>
-                  <Box sx={{ display: 'flex' }}>
-                    <Typography variant="body2">
-                      {wallet.name}
-                    </Typography>
-                  </Box>
-                </TableCell>
-                <TableCell>
-                  <Box sx={{ display: 'flex' }}>
-                    <Typography variant="body2" noWrap>
-                      {wallet.address}
-                    </Typography>
-                    <CopyButton onClick={() => handleCopyToClipboard(wallet.address)}>
-                      Copy
-                    </CopyButton>
-                  </Box>
-                </TableCell>
-                <TableCell>
-                  <Box>
-                    <a
-                      href={wallet.arbiscanLink}
-                      target="_blank"
-                      rel="noopener noreferrer"
-                      style={{ cursor: 'pointer', marginRight: 10 }}
-                    >
-                      <img
-                        src={arbitrumLogo}
-                        alt="arbitrum icon"
-                        width={30}
-                      />
-                    </a>
-                  </Box>
-                </TableCell>
-              </TableRow>
-            ))}
-          </TableBody>
-        </Table>
-      </TableContainer>
-    </ThemeProvider>
+          ))}
+        </TableBody>
+      </Table>
+    </TableContainer>
   );
 };
