+++ apps/web/src/theme/index.ts
@@ -1,62 +1,83 @@
-export interface FontSizes {
-  h1: string;
-  h2: string;
-  h3: string;
-  h4: string;
-  h5: string;
-  h6: string;
-  button: string;
-  body: string;
-  subtitle: string;
-  p: string;
-  caption: string;
+import { createTheme, ThemeOptions } from '@mui/material/styles';
+import { TypographyOptions } from "@mui/material/styles/createTypography";
 
-  xs: string;
-  sm: string;
-  base: string;
-  lg: string;
-  xl: string;
-  '2xl': string;
-  '3xl': string;
-  '4xl': string;
-  '5xl': string;
-  '6xl': string;
-  '7xl': string;
-  '8xl': string;
-  '9xl': string;
+import ApercuMonoProBold from './fonts/ApercuMonoProBold.ttf';
+import ApercuMonoProMedium from './fonts/ApercuMonoProMedium.ttf';
+import ApercuMonoProRegular from './fonts/ApercuMonoProRegular.ttf';
+import ApercuMonoProLight from './fonts/ApercuMonoProLight.ttf';
+import AvenirNextBold from './fonts/AvenirNextLTPro-Bold.otf';
+import AvenirNextRegular from './fonts/AvenirNextLTPro-Regular.otf';
+import AvenirNextLite from './fonts/AvenirNextLTPro-It.otf';
+
+import '@mui/material/styles';
+
+
+declare module '@mui/material/styles' {
+  interface TypographyVariants {
+    xs: React.CSSProperties,
+    sm: React.CSSProperties,
+    base: React.CSSProperties,
+    lg: React.CSSProperties,
+    xl: React.CSSProperties,
+    '2xl': React.CSSProperties,
+    '3xl': React.CSSProperties,
+    '4xl': React.CSSProperties,
+    '5xl': React.CSSProperties,
+    '6xl': React.CSSProperties,
+    '7xl': React.CSSProperties,
+    '8xl': React.CSSProperties,
+    '9xl': React.CSSProperties,
+  }
+
+  interface TypographyVariantsOptions {
+    xs: React.CSSProperties,
+    sm: React.CSSProperties,
+    base: React.CSSProperties,
+    lg: React.CSSProperties,
+    xl: React.CSSProperties,
+    '2xl': React.CSSProperties,
+    '3xl': React.CSSProperties,
+    '4xl': React.CSSProperties,
+    '5xl': React.CSSProperties,
+    '6xl': React.CSSProperties,
+    '7xl': React.CSSProperties,
+    '8xl': React.CSSProperties,
+    '9xl': React.CSSProperties,
+  }
 }
 
-export interface LineHeights {
-  h1: string;
-  h2: string;
-  h3: string;
-  h4: string;
-  h5: string;
-  h6: string;
-  button: string;
-  body: string;
-  subtitle: string;
-  p: string;
-  caption: string;
-
-  xs: string;
-  sm: string;
-  base: string;
-  lg: string;
-  xl: string;
-  '2xl': string;
-  '3xl': string;
-  '4xl': string;
-  '5xl': string;
-  '6xl': string;
-  '7xl': string;
-  '8xl': string;
-  '9xl': string;
+declare module '@mui/material/Typography' {
+  interface TypographyPropsVariantOverrides {
+    xs: true,
+    sm: true,
+    base: true,
+    lg: true,
+    xl: true,
+    '2xl': true,
+    '3xl': true,
+    '4xl': true,
+    '5xl': true,
+    '6xl': true,
+    '7xl': true,
+    '8xl': true,
+    '9xl': true,
+  }
 }
 
-export interface FontFamilies {
-  primary: string;
-  secondary: string;
+interface ExtendedTypographyOptions extends TypographyOptions {
+  xs: React.CSSProperties,
+  sm: React.CSSProperties,
+  base: React.CSSProperties,
+  lg: React.CSSProperties,
+  xl: React.CSSProperties,
+  '2xl': React.CSSProperties,
+  '3xl': React.CSSProperties,
+  '4xl': React.CSSProperties,
+  '5xl': React.CSSProperties,
+  '6xl': React.CSSProperties,
+  '7xl': React.CSSProperties,
+  '8xl': React.CSSProperties,
+  '9xl': React.CSSProperties,
 }
 
 export interface Colors {
@@ -88,74 +109,6 @@ export interface Colors {
   verticalWhiteToTransparent: string
 }
 
-export interface Theme {
-  fontSizes: FontSizes,
-  fontFamilies: FontFamilies,
-  lineHeights: LineHeights,
-  colors: Colors,
-}
-
-const fontSizes: FontSizes = {
-  h1: '104px',
-  h2: '96px',
-  h3: '75px',
-  h4: '48px',
-  h5: '40px',
-  h6: '32px',
-  button: '32px',
-  body: '28px',
-  subtitle: '24px',
-  p: '18px',
-  caption: '16px',
-
-  xs: '0.75rem',
-  sm: '0.875rem',
-  base: '1rem',
-  lg: '1.125rem',
-  xl: '1.25rem',
-  '2xl': '1.5rem',
-  '3xl': '1.875rem',
-  '4xl': '2.25rem',
-  '5xl': '3rem',
-  '6xl': '3.75rem',
-  '7xl': '4.5rem',
-  '8xl': '5rem',
-  '9xl': '8rem',
-};
-
-const lineHeights: LineHeights = {
-  h1: '116px',
-  h2: '116px',
-  h3: '108px',
-  h4: '72px',
-  h5: '60px',
-  h6: '48px',
-  button: '140px',
-  body: '42px',
-  subtitle: '42px',
-  p: '26px',
-  caption: '24px',
-
-  xs: '1rem',
-  sm: '1.25rem',
-  base: '1.5rem',
-  lg: '1.75rem',
-  xl: '1.75rem',
-  '2xl': '2rem',
-  '3xl': '2.25rem',
-  '4xl': '2.5rem',
-  '5xl': '1rem',
-  '6xl': '1rem',
-  '7xl': '1rem',
-  '8xl': '1rem',
-  '9xl': '1rem',
-};
-
-const fontFamilies: FontFamilies = {
-  primary: 'Apercu Mono Pro',
-  secondary: 'Avenir Next',
-};
-
 const colors: Colors = {
   dark: '#0A0A0A',
   white: '#FFFFFF',
@@ -185,11 +138,155 @@ const colors: Colors = {
   verticalWhiteToTransparent: 'linear-gradient(180deg, rgba(255, 255, 255, 0.8) 7.41%, rgba(255, 255, 255, 0) 85.19%)',
 };
 
-const theme = {
-  fontSizes,
-  lineHeights,
-  fontFamilies,
-  colors,
-};
+const theme = createTheme({
+  breakpoints: { 
+    values: {
+      xs: 0,     // 0em
+      sm: 640,   // 40em
+      md: 832,   // 52em
+      lg: 1024,  // 64em
+      xl: 1280   // 80em
+    }
+  },
+  components: {
+    MuiCssBaseline: {
+      styleOverrides: `
+        @font-face {
+          font-family: "Apercu Mono Pro";
+          src: url(${ApercuMonoProBold}) format("truetype");
+          font-weight: 700;
+        }
+
+        @font-face {
+          font-family: "Apercu Mono Pro";
+          src: url(${ApercuMonoProMedium}) format("truetype");
+          font-weight: 300;
+        }
+
+        @font-face {
+          font-family: "Apercu Mono Pro";
+          src: url(${ApercuMonoProRegular}) format("truetype");
+          font-weight: 500;
+        }
+
+        @font-face {
+          font-family: "Apercu Mono Pro";
+          src: url(${ApercuMonoProLight}) format("truetype");
+          font-weight: 400;
+        }
+
+        @font-face {
+          font-family: "Avenir Next";
+          src: url(${AvenirNextBold}) format("opentype");
+          font-weight: 700;
+        }
+
+        @font-face {
+          font-family: "Avenir Next";
+          src: url(${AvenirNextRegular}) format("opentype");
+          font-weight: 500;
+        }
+
+        @font-face {
+          font-family: "Avenir Next";
+          src: url(${AvenirNextLite}) format("opentype");
+          font-weight: 400;
+        }
+
+        body {
+          min-height: 100vh;
+          margin: 0;
+          color: white;
+          background-color: #0A0A0A;
+          -webkit-font-smoothing: antialiased;
+          -moz-osx-font-smoothing: grayscale;
+        }
+
+        a, a:visited {
+          color: #09FEF5;
+          text-decoration: none;
+        }
+
+        *:focus,
+        button:focus {
+          outline: none;
+        }
+      `,
+    },
+  },
+  palette: {
+    mode: 'dark',
+    common: {
+      black: '#0A0A0A',
+      white: '#FFFFFF'
+    },
+    primary: {
+      main: '#B6FE46',
+    },
+    secondary: {
+      main: '#09FEF5',
+    },
+    error: {
+      main: '#fe092e',
+    },
+    grey: {
+      50: '#F8F9FA',
+      100: '#EBEDEF',
+      200: '#DDE1E4',
+      300: '#CED3D8',
+      400: '#BDC4CB',
+      500: '#AAB4BC',
+      600: '#95A1AC',
+      700: '#808A93',
+      800: '#646D74',
+      900: '#3B3F44',
+    },
+    background: {
+      default: '#121212',
+      paper: '#1D1D1D',
+    },
+    text: {
+      primary: colors.white,
+      secondary: colors['white-80'],
+      disabled: 'rgba(255, 255, 255, 0.38)',
+    },
+    action: {
+      hover: 'rgba(255, 255, 255, 0.08)',
+      selected: 'rgba(255, 255, 255, 0.16)',
+      disabled: 'rgba(255, 255, 255, 0.38)',
+      disabledBackground: 'rgba(255, 255, 255, 0.12)',
+    },
+    horizontalGreenToCyan: 'linear-gradient(94.56deg, #B6FE46 7.15%, rgba(9, 254, 245, 0.6) 90.05%)',
+    verticalGreenToTransparent: 'linear-gradient(180deg, rgba(182, 254, 70, 0.8) 0%, rgba(217, 217, 217, 0) 100%)',
+    verticalWhiteToTransparent: 'linear-gradient(180deg, rgba(255, 255, 255, 0.8) 7.41%, rgba(255, 255, 255, 0) 85.19%)',
+  },
+  typography: {
+    fontFamily: '"Apercu Mono Pro", "Avenir Next", sans-serif',
+    h1: { fontSize: '104px', lineHeight: '116px' },
+    h2: { fontSize: '96px', lineHeight: '116px' },
+    h3: { fontSize: '75px', lineHeight: '108px' },
+    h4: { fontSize: '48px', lineHeight: '72px' },
+    h5: { fontSize: '40px', lineHeight: '60px' },
+    h6: { fontSize: '32px', lineHeight: '48px' },
+    button: { fontSize: '32px', lineHeight: '140px' },
+    body1: { fontSize: '28px', lineHeight: '42px' },
+    body2: { fontSize: '18px', lineHeight: '26px' },      // p
+    subtitle1: { fontSize: '24px', lineHeight: '42px' },  // subtitle
+    caption: { fontSize: '16px', lineHeight: '24px' },
+    xs: { fontSize: '0.75rem', lineHeight: '1rem' },
+    sm: { fontSize: '0.875rem', lineHeight: '1.25rem' },
+    base: { fontSize: '1rem', lineHeight: '1.5rem' },
+    lg: { fontSize: '1.125rem', lineHeight: '1.75rem' },
+    xl: { fontSize: '1.25rem', lineHeight: '1.75rem' },
+    '2xl': { fontSize: '1.5rem', lineHeight: '2rem' },
+    '3xl': { fontSize: '1.875rem', lineHeight: '2.25rem', fontFamily: "Avenir Next" },
+    '4xl': { fontSize: '2.25rem', lineHeight: '2.5rem' },
+    '5xl': { fontSize: '3rem', lineHeight: '3rem', fontFamily: "Avenir Next" },
+    '6xl': { fontSize: '3.75rem', lineHeight: '4rem' },
+    '7xl': { fontSize: '4.5rem', lineHeight: '5.5rem', fontFamily: "Avenir Next" },
+    '8xl': { fontSize: '5rem', lineHeight: '6.2rem', fontFamily: "Avenir Next" },
+    '9xl': { fontSize: '6.5rem', lineHeight: '7.5rem', fontFamily: "Avenir Next" }
+  } as ExtendedTypographyOptions,
+} as ThemeOptions);
 
 export default theme;
