--- apps/web/src/theme/styled.d.ts
@@ -1,7 +1,0 @@
-import 'styled-components';
-import { Theme } from './';
-
-declare module 'styled-components' {
-  export interface DefaultTheme extends Theme {
-  }
-}
